% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tblout.R
\name{as_tblout}
\alias{as_tblout}
\title{Tabular output format for a scan}
\usage{
as_tblout(x, output = "", append = FALSE)
}
\arguments{
\item{x}{A \link{data.frame} as returned by \code{\link[=pfamscan]{pfamscan()}}.}

\item{output}{Character scalar. Output file where to write the table}

\item{append}{Passed to \code{\link[=cat]{cat()}}.}
}
\value{
Prints the file on screen and returns the tabular format as a \link{data.frame}
invisible.
}
\description{
This resembles the behavior of HMM when the option \code{--tblout} is passed.
The resulting file is space separated with varying widths, so the user
should not rely on the file to be a fixed-width-file.
}
\examples{
data("P35858", package = "pfamscanr")
as_tblout(P35858)
}
