% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfamscanr-package.R
\docType{data}
\name{P35858}
\alias{P35858}
\title{Insulin-like growth factor-binding protein complex acid labile subunit}
\format{\describe{
\item{\code{model_length}}{Character}
\item{\code{align}}{List}
\item{\code{env}}{data.frame with 2 variables}
\item{\code{name}}{Character}
\item{\code{acc}}{Character}
\item{\code{sig}}{Integer}
\item{\code{evalue}}{Character}
\item{\code{desc}}{Character}
\item{\code{hmm}}{data.frame with two columns}
\item{\code{act_site}}{logical}
\item{\code{type}}{character}
\item{\code{bits}}{character}
\item{\code{clan}}{character}
\item{\code{seq}}{data.frame with three variables}
}}
\source{
\url{https://www.uniprot.org/uniprot/P35858}
}
\usage{
P35858
}
\description{
A dataset containing the result from querying:
\preformatted{>sp|P35858|ALS_HUMAN ... OS=Homo sapiens GN=IGFALS PE=1 SV=1
MALRKGGLALALLLLSWVALGPRSLEGADPGTPGEAEGPACPAACVCSYDD
DADELSVFCSSRNLTRLPDGVPGGTQALWLDGNNLSSVPPAAFQNLSSLGF
LNLQGGQLGSLEPQALLGLENLCHLHLERNQLRSLALGTFAHTPALASLGL
SNNRLSRLEDGLFEGLGSLWDLNLGWNSLAVLPDAAFRGLGSLRELVLAGN
RLAYLQPALFSGLAELRELDLSRNALRAIKANVFVQLPRLQKLYLDRNLIA
AVAPGAFLGLKALRWLDLSHNRVAGLLEDTFPGLLGLRVLRLSHNAIASLR
PRTFKDLHFLEELQLGHNRIRQLAERSFEGLGQLEVLTLDHNQLQEVKAGA
FLGLTNVAVMNLSGNCLRNLPEQVFRGLGKLHSLHLEGSCLGRIRPHTFTG
LSGLRRLFLKDNGLVGIEEQSLWGLAELLELDLTSNQLTHLPHRLFQGLGK
LEYLLLSRNRLAELPADALGPLQRAFWLDVSHNRLEALPNSLLAPLGRLRY
LSLRNNSLRTFTPQPPGLERLWLEGNPWDCGCPLKALRDFALQNPSAVPRF
VQAICEGDDCQPPAYTYNNITCASPPEVVGLDLRDLSEAHFAPC}
}
\keyword{datasets}
