% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfamscan.R
\name{pfamscan}
\alias{pfamscan}
\title{Query the online service PfamScan}
\usage{
pfamscan(
  fasta_str,
  email,
  ...,
  api_url = "https://www.ebi.ac.uk/",
  path = "Tools/services/rest/pfamscan",
  maxchecktime = 120,
  wait = 1,
  errorfun = message,
  max_post_size = 2000L,
  verb = TRUE
)
}
\arguments{
\item{fasta_str}{A string representing a single sequence in fasta format}

\item{email}{Required by the API.}

\item{...}{Further arguments passed to \code{\link[=httr:POST]{httr:POST()}} and \code{\link[=httr:GET]{httr:GET()}}.}

\item{api_url, path}{URL path to the API.}

\item{maxchecktime}{Max wait time for the results (seconds).}

\item{wait}{Time, in seconds, to wait between checks.}

\item{errorfun}{A function to either print error or message.}

\item{max_post_size}{Maximum length of the string to POST using curl.}

\item{verb}{Logical scalar. When \code{TRUE} the function will print informative
messages.}
}
\description{
This uses the EBI API. Details here:
https://www.ebi.ac.uk/Tools/common/tools/help/index.html
}
\references{
Madeira F, Park YM, Lee J, et al. The EMBL-EBI search and sequence analysis
tools APIs in 2019. Nucleic Acids Research. 2019 Jul;47(W1):W636-W641.
\doi{10.1093/nar/gkz268}.
}
