% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{data}
\name{pfamscanr_queries}
\alias{pfamscanr_queries}
\title{Queries performed during the current R Session}
\format{
An object of class \code{pfamscanr_queries} (inherits from \code{environment}) of length 4.
}
\usage{
pfamscanr_queries
}
\description{
Queries performed during the current R Session
}
\keyword{datasets}
