% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfamscan.R
\name{pfamscan_retrieve}
\alias{pfamscan_retrieve}
\alias{pfamscan}
\title{Query the online service PfamScan}
\usage{
pfamscan_retrieve(query_id, ..., api_url = "https://www.ebi.ac.uk/",
  path = "Tools/services/rest/pfamscan", errorfun = message,
  verb = TRUE, json_args = list())

pfamscan(fasta_str, email, ..., evalue = 10, asp = "false",
  api_url = "https://www.ebi.ac.uk/",
  path = "Tools/services/rest/pfamscan", maxchecktime = 120,
  wait = 1, errorfun = message, query_size = 100L, verb = TRUE,
  json_args = list())
}
\arguments{
\item{query_id}{Character scalar. Id of the query to retrieve, e.g.,
\code{"pfamscan-R20191015-22133-0197-415408-p2m"}}

\item{...}{Further arguments passed to \code{\link[httr:POST]{httr::POST()}} and \code{\link[httr:GET]{httr::GET()}}.}

\item{api_url, path}{URL path to the API.}

\item{errorfun}{A function to either print error or message.}

\item{verb}{Logical scalar. When \code{TRUE} the function will print informative
messages.}

\item{json_args}{Arguments passed to \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}} (see details).}

\item{fasta_str}{A string representing a single sequence in fasta format (details)}

\item{email, evalue, asp}{Query parameters, see \href{https://www.ebi.ac.uk/seqdb/confluence/display/THD/PfamScan}{here}.}

\item{maxchecktime}{Max wait time for the results (seconds).}

\item{wait}{Time, in seconds, to wait between checks.}

\item{query_size}{Maximum number of sequences per query (limit 100).}
}
\description{
This uses the EBI API. Details here:
https://www.ebi.ac.uk/Tools/common/tools/help/index.html
}
\details{
The function \code{pfamscan_retrieve} can be used to retrieve the results of
an specific query. This is useful if, for example, the query was correctly
posted but the function ran out of time waiting for it to be completed.
You can check the list of posted queries during the session using
\link{pfamscanr_queries}

If \code{fasta_str} is a vector including multiple sequences to submit, the
function will make either a single call to the API or multiple depending
on \code{query_size}. As of now, the API limits 100 sequences per query.

The function returns the response (in JSON format) parsed using the function
\code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}.
}
\examples{
\dontrun{
# This is an example extracted directly from the website
# https://www.ebi.ac.uk/Tools/pfa/pfamscan/
ans <- pfamscan(
  fasta_str = ">sp|P35858|ALS_HUMAN Insulin-like growth factor-binding protein complex acid labile subunit OS=Homo sapiens GN=IGFALS PE=1 SV=1
MALRKGGLALALLLLSWVALGPRSLEGADPGTPGEAEGPACPAACVCSYDD
DADELSVFCSSRNLTRLPDGVPGGTQALWLDGNNLSSVPPAAFQNLSSLGF
LNLQGGQLGSLEPQALLGLENLCHLHLERNQLRSLALGTFAHTPALASLGL
SNNRLSRLEDGLFEGLGSLWDLNLGWNSLAVLPDAAFRGLGSLRELVLAGN
RLAYLQPALFSGLAELRELDLSRNALRAIKANVFVQLPRLQKLYLDRNLIA
AVAPGAFLGLKALRWLDLSHNRVAGLLEDTFPGLLGLRVLRLSHNAIASLR
PRTFKDLHFLEELQLGHNRIRQLAERSFEGLGQLEVLTLDHNQLQEVKAGA
FLGLTNVAVMNLSGNCLRNLPEQVFRGLGKLHSLHLEGSCLGRIRPHTFTG
LSGLRRLFLKDNGLVGIEEQSLWGLAELLELDLTSNQLTHLPHRLFQGLGK
LEYLLLSRNRLAELPADALGPLQRAFWLDVSHNRLEALPNSLLAPLGRLRY
LSLRNNSLRTFTPQPPGLERLWLEGNPWDCGCPLKALRDFALQNPSAVPRF
VQAICEGDDCQPPAYTYNNITCASPPEVVGLDLRDLSEAHFAPC",
  email     = "your@email.com",
  httr::config(connecttimeout=60)
)
}
}
\references{
Madeira F, Park YM, Lee J, et al. The EMBL-EBI search and sequence analysis
tools APIs in 2019. Nucleic Acids Research. 2019 Jul;47(W1):W636-W641.
\doi{10.1093/nar/gkz268}.
}
