% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\docType{data}
\name{pfamscanr_queries}
\alias{pfamscanr_queries}
\title{Queries performed during the current R Session}
\format{
An object of class \code{pfamscanr_queries} (inherits from \code{environment}) of length 4.
}
\usage{
pfamscanr_queries
}
\description{
This environment keeps track of the query ids successfully submitted
and whether they were marked as "done" or not.
}
\details{
The \link{print} method prints a data.frame with the information.
The data is stored directly in the \code{dat} slot.
}
\keyword{datasets}
